# used deque instead of list in ipreorder